refactor: move dal errors into backend/dal
We should limit our top-level packages.
