+++ .gitignore
@@ -24,6 +24,5 @@ buildengine/.gitignore
 go.work*
 junit*.xml
 /readme-tests
-**/_ftl
 /docs/public
 .ftl.lock
