+++ backend/controller/dal/pubsub.go
@@ -6,8 +6,8 @@ import (
 	"time"
 
 	"github.com/TBD54566975/ftl/backend/controller/sql"
+	dalerrs "github.com/TBD54566975/ftl/backend/dal"
 	"github.com/TBD54566975/ftl/backend/schema"
-	"github.com/TBD54566975/ftl/db/dalerrs"
 	"github.com/TBD54566975/ftl/internal/log"
 	"github.com/TBD54566975/ftl/internal/model"
 	"github.com/TBD54566975/ftl/internal/slices"
